+++ changelog.md
@@ -1,9 +1,14 @@
+# v1.4.0
+* Added Custom Backgrounds!
+  * You can add backgrounds with the same process as stickers
+* Cleaned up settings
+* Added more background customization settings.
+  * You can now change the position and scale of your background
 # v1.3.0
 * Added Custom Stickers!
   * You can add the path to a sticker in settings
   * Enable the custom stickers setting to use
 * Readded MacOS Support
-
 # v1.2.0
 * Ported to GD 2.206 and Geode 3.4.0
 * Fixed a bug with editor backgrounds going off screen